Move concluded integrations to the new listener
So that all listeners to the same event are in the same place
